+++ impeller/BUILD.gn
@@ -66,7 +66,6 @@ group("impeller") {
       "aiks",
       "display_list",
       "entity",
-      "image:image_skia_backend",
       "renderer",
       "renderer/backend",
       "typographer/backends/skia:typographer_skia_backend",
@@ -99,7 +98,6 @@ impeller_component("impeller_unittests") {
       "entity:entity_unittests",
       "fixtures",
       "geometry:geometry_unittests",
-      "image:image_unittests",
       "playground",
       "renderer:renderer_unittests",
       "scene:scene_unittests",
