+++ impeller/playground/image/compressed_image.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "impeller/image/compressed_image.h"
+#include "impeller/playground/image/compressed_image.h"
 
 namespace impeller {
