+++ impeller/playground/image/decompressed_image.h
@@ -2,8 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#ifndef FLUTTER_IMPELLER_IMAGE_DECOMPRESSED_IMAGE_H_
-#define FLUTTER_IMPELLER_IMAGE_DECOMPRESSED_IMAGE_H_
+#ifndef FLUTTER_IMPELLER_PLAYGROUND_IMAGE_DECOMPRESSED_IMAGE_H_
+#define FLUTTER_IMPELLER_PLAYGROUND_IMAGE_DECOMPRESSED_IMAGE_H_
 
 #include <memory>
 #include <optional>
@@ -51,4 +51,4 @@ class DecompressedImage {
 
 }  // namespace impeller
 
-#endif  // FLUTTER_IMPELLER_IMAGE_DECOMPRESSED_IMAGE_H_
+#endif  // FLUTTER_IMPELLER_PLAYGROUND_IMAGE_DECOMPRESSED_IMAGE_H_
